feat: Refactor codebase.
