+++ .github/workflows/ci.yaml
@@ -1,4 +1,4 @@
-name: ecgarr
+name: heartkit
 
 on: [push]
