+++ configs/download-datasets.json
@@ -1,5 +1,5 @@
 {
-    "db_root_path": "./datasets",
+    "ds_path": "./datasets",
     "datasets": ["icentia11k", "ludb"],
     "progress": true
 }
