--- ecgarr/__main__.py
@@ -1,3 +1,0 @@
-from ecgarr import app
-
-app.run()
