+++ heartkit/__main__.py
@@ -1,0 +1,3 @@
+from heartkit import app
+
+app.run()
