+++ heartkit/app.py
@@ -5,13 +5,13 @@ import pydantic_argparse
 from pydantic import BaseModel, Field
 
 from . import datasets as ds
-from . import demo, deploy, evaluate, train
+from . import demo, evaluate, export, train
 from .types import (
-    EcgDemoParams,
-    EcgDeployParams,
-    EcgDownloadParams,
-    EcgTestParams,
-    EcgTrainParams,
+    HeartDemoParams,
+    HeartDownloadParams,
+    HeartExportParams,
+    HeartTestParams,
+    HeartTrainParams,
 )
 from .utils import setup_logger
 
@@ -27,11 +27,11 @@ class AppCommandArgments(BaseModel):
 class AppArguments(BaseModel):
     """App CLI arguments"""
 
-    download_dataset: Optional[AppCommandArgments] = Field(description="Fetch dataset")
-    train_model: Optional[AppCommandArgments] = Field(description="Train model")
-    test_model: Optional[AppCommandArgments] = Field(description="Test model")
-    deploy_model: Optional[AppCommandArgments] = Field(description="Deploy model")
-    evb_demo: Optional[AppCommandArgments] = Field(description="EVB demo")
+    download: Optional[AppCommandArgments] = Field(description="Fetch dataset")
+    train: Optional[AppCommandArgments] = Field(description="Train model")
+    evaluate: Optional[AppCommandArgments] = Field(description="Test model")
+    export: Optional[AppCommandArgments] = Field(description="Export model")
+    demo: Optional[AppCommandArgments] = Field(description="EVB demo")
 
 
 def download_dataset(command: AppCommandArgments):
@@ -39,7 +39,7 @@ def download_dataset(command: AppCommandArgments):
     Args:
         command (AppCommandArgments): Command arguments
     """
-    params = EcgDownloadParams.parse_file(command.config_file)
+    params = HeartDownloadParams.parse_file(command.config_file)
     logger.info("#STARTED downloading dataset")
     ds.download_datasets(params=params)
     logger.info("#FINISHED downloading dataset")
@@ -50,32 +50,32 @@ def train_model(command: AppCommandArgments):
     Args:
         command (AppCommandArgments): Command arguments
     """
-    params = EcgTrainParams.parse_file(command.config_file)
+    params = HeartTrainParams.parse_file(command.config_file)
     logger.info("#STARTED training model")
     train.train_model(params=params)
     logger.info("#FINISHED training model")
 
 
-def test_model(command: AppCommandArgments):
+def evaluate_model(command: AppCommandArgments):
     """Test model CLI command.
     Args:
         command (AppCommandArgments): Command arguments
     """
 
-    params = EcgTestParams.parse_file(command.config_file)
+    params = HeartTestParams.parse_file(command.config_file)
     logger.info("#STARTED testing model")
     evaluate.evaluate_model(params=params)
     logger.info("#FINISHED testing model")
 
 
-def deploy_model(command: AppCommandArgments):
+def export_model(command: AppCommandArgments):
     """Deploy model CLI command.
     Args:
         command (AppCommandArgments): Command arguments
     """
-    params = EcgDeployParams.parse_file(command.config_file)
+    params = HeartExportParams.parse_file(command.config_file)
     logger.info("#STARTED deploying model")
-    deploy.deploy_model(params=params)
+    export.export_model(params=params)
     logger.info("#FINISHED deploying model")
 
 
@@ -84,7 +84,7 @@ def evb_demo(command: AppCommandArgments):
     Args:
         command (AppCommandArgments): Command arguments
     """
-    params = EcgDemoParams.parse_file(command.config_file)
+    params = HeartDemoParams.parse_file(command.config_file)
     logger.info("#STARTED evb demo")
     demo.evb_demo(params=params)
     logger.info("#FINISHED evb demo")
@@ -97,21 +97,21 @@ def run(inputs: Optional[List[str]] = None):
     """
     parser = pydantic_argparse.ArgumentParser(
         model=AppArguments,
-        prog="ECG Heart Demo",
-        description="ECG Heart demo",
+        prog="Heart Kit Demo",
+        description="Heart Kit demo",
     )
     args = parser.parse_typed_args(inputs)
 
-    if args.download_dataset:
-        download_dataset(args.download_dataset)
-    elif args.train_model:
-        train_model(args.train_model)
-    elif args.test_model:
-        test_model(args.test_model)
-    elif args.deploy_model:
-        deploy_model(args.deploy_model)
-    elif args.evb_demo:
-        evb_demo(args.evb_demo)
+    if args.download:
+        download_dataset(args.download)
+    elif args.train:
+        train_model(args.train)
+    elif args.evaluate:
+        evaluate_model(args.evaluate)
+    elif args.export:
+        export_model(args.export)
+    elif args.demo:
+        evb_demo(args.demo)
     else:
         logger.error("Error: Unknown command")
