+++ heartkit/datasets/download.py
@@ -2,27 +2,27 @@ import os
 
 import pydantic_argparse
 
-from ..types import EcgDownloadParams
+from ..types import HeartDownloadParams
 from .icentia11k import IcentiaDataset
 from .ludb import LudbDataset
 
 
-def download_datasets(params: EcgDownloadParams):
+def download_datasets(params: HeartDownloadParams):
     """Download all specified datasets.
 
     Args:
-        params (EcgDownloadParams): Download parameters
+        params (HeartDownloadParams): Download parameters
     """
-    os.makedirs(params.ds_root_path, exist_ok=True)
-    #### Icentia11k Dataset
+    os.makedirs(params.ds_path, exist_ok=True)
+
     if "icentia11k" in params.datasets:
-        IcentiaDataset(str(params.ds_root_path)).download(
+        IcentiaDataset(str(params.ds_path)).download(
             num_workers=params.data_parallelism,
             force=params.force,
         )
 
     if "ludb" in params.datasets:
-        LudbDataset(str(params.ds_root_path)).download(
+        LudbDataset(str(params.ds_path)).download(
             num_workers=params.data_parallelism,
             force=params.force,
         )
@@ -31,7 +31,7 @@ def download_datasets(params: EcgDownloadParams):
 def create_parser():
     """Create CLI parser"""
     return pydantic_argparse.ArgumentParser(
-        model=EcgDownloadParams,
+        model=HeartDownloadParams,
         prog="ECG Dataset",
         description="ECG dataset",
     )
