+++ heartkit/tasks/__init__.py
@@ -1,0 +1,54 @@
+from typing import Optional, Tuple
+
+import tensorflow as tf
+from keras.engine.keras_tensor import KerasTensor
+
+from ..models.features import ecg_feature_extractor
+from ..types import ArchitectureType, HeartTask, get_num_classes
+
+
+def create_task_model(
+    inputs: KerasTensor,
+    task: HeartTask,
+    arch: ArchitectureType = "resnet18",
+    stages: Optional[int] = None,
+) -> tf.keras.Model:
+    """Generate model for given task
+
+    Args:
+        inputs (KerasTensor): Model inputs
+        task (HeartTask): Heart task
+        arch (ArchitectureType, optional): Architecture type. Defaults to 'resnet18'.
+        stages (Optional[int], optional): # stages in network. Defaults to None.
+
+    Returns:
+        tf.keras.Model: Model
+    """
+    num_classes = get_num_classes(task=task)
+    x = ecg_feature_extractor(inputs, arch, stages=stages)
+    outputs = tf.keras.layers.Dense(num_classes)(x)
+    model = tf.keras.Model(inputs, outputs, name="model")
+    return model
+
+
+def get_task_spec(task: HeartTask, frame_size: int) -> Tuple[tf.TensorSpec]:
+    """Get task model spec
+
+    Args:
+        task (HeartTask): ECG task
+        frame_size (int): Frame size
+
+    Returns:
+        Tuple[tf.TensorSpec]: TF spec for task
+    """
+    if task in [HeartTask.rhythm, HeartTask.beat, HeartTask.hr]:
+        return (
+            tf.TensorSpec((frame_size, 1), tf.float32),
+            tf.TensorSpec((), tf.int32),
+        )
+    if task in [HeartTask.segmentation]:
+        return (
+            tf.TensorSpec((frame_size, 1), tf.float32),
+            tf.TensorSpec((frame_size, 1), tf.int32),
+        )
+    raise ValueError(f"unknown task: {task}")
